+++ src/main/java/com/dt/user/service/RoleMenuService.java
@@ -10,7 +10,7 @@ public interface RoleMenuService {
     /**
      * 通过角色id来查询拥有的菜单
      **/
-    List<String> gerRoleMenus(Long rid);
+    List<String> gerRoleMenus(Integer rid);
 
     /**
      * 添加角色菜单/删除角色菜单
